backend/requirements.txt
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
Django==3.2.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
geographiclib==1.52
geopy==2.2.0
h11==0.13.0
idna==3.3
outcome==1.1.0
psycopg2-binary==2.9.3
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
pytz==2022.1
requests==2.27.1
selenium==4.1.2
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
webdriver-manager==3.5.4
wsproto==1.1.0
